requirements.txt
fastapi
uvicorn[standard]
pillow
numpy
torch
torchvision